requirements.txt
alembic==1.4.1
astroid==2.3.3
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.9
django-nexmo==1.0.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-ngrok==0.0.25
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pyotp==2.3.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-qt==0.50
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.15
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
